moved sample files, options for more input on project creation
