+++ src/workflows/spark-mpi/components/root/NewSimulation.js
@@ -1,7 +1,7 @@
 import React                from 'react';
 import { FileUploadEntry }  from '../../../../panels/ItemEditor';
 
-const newSim = (props) => <FileUploadEntry name="dataset" label="Tif" owner={props.owner} />;
+const newSim = (props) => <FileUploadEntry name="input" label="Input file" accept="*" owner={props.owner} />;
 
 newSim.propTypes = {
   owner: React.PropTypes.func,
